Put drawing and calculations in one thread
