+++ app/src/main/java/oug/com/pong/GameActivity.java
@@ -5,10 +5,14 @@ import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.Display;
+import android.view.MotionEvent;
+import android.view.View;
 
-public class GameActivity extends AppCompatActivity {
+public class GameActivity extends AppCompatActivity implements View.OnTouchListener, Runnable{
     GameView gameView;
     GameModel gameModel;
+    Thread gameThread = null;
+    boolean isRunning=false;
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -18,22 +22,53 @@ public class GameActivity extends AppCompatActivity {
         display.getSize(size);
         int width = size.x;
         int height = size.y;
-
         gameModel = new GameModel(width,height);
         gameView = new GameView(this,gameModel);
+        gameView.setOnTouchListener(this);
         setContentView(gameView);
+
+        gameModel.setSpeed(7);
+        gameModel.newGame();
 
     }
 
     @Override
     protected void onPause() {
         super.onPause();
-        gameView.pause();
+        isRunning=false;
+        while(true){
+            try {
+                gameThread.join();
+            } catch (InterruptedException e) {
+                e.printStackTrace();
+            }
+            break;
+        }
+        gameThread=null;
     }
 
     @Override
     protected void onResume() {
         super.onResume();
-        gameView.resume();
+        isRunning=true;
+        gameThread = new Thread(this);
+        gameThread.start();
+    }
+
+    @Override
+    public boolean onTouch(View view, MotionEvent motionEvent) {
+        gameModel.setPaddleBY((int) motionEvent.getY());
+        return true;
+    }
+
+    @Override
+    public void run() {
+        while(isRunning){
+            Log.i("testing",gameModel.getBallX() + ","+gameModel.getBallY());
+            gameModel.moveBall();
+            gameModel.AI();
+            gameModel.checkCollisions();
+            gameView.draw();
+        }
     }
 }
