+++ app/src/main/java/oug/com/pong/GameView.java
@@ -9,10 +9,8 @@ import android.util.Log;
 import android.view.SurfaceHolder;
 import android.view.SurfaceView;
 
-public class GameView extends SurfaceView implements Runnable{
+public class GameView extends SurfaceView{
     SurfaceHolder holder;
-    Thread gameThread = null;
-    boolean isRunning=false;
     GameModel gameModel;
 
     public GameView(Context context, GameModel gameModel) {
@@ -21,43 +19,21 @@ public class GameView extends SurfaceView implements Runnable{
         this.gameModel=gameModel;
 
     }
-
-    @Override
-    public void run() {
-        while(isRunning){
-            if(!holder.getSurface().isValid()){
-                continue;
-            }
-            Paint white = new Paint();
-            white.setStyle(Paint.Style.FILL);
-            white.setColor(Color.WHITE);
-            Canvas c = holder.lockCanvas();
-            c.drawARGB(255,0,0,0);
-            c.drawRect(0,gameModel.getPaddleAY(), gameModel.getPaddleWidth(),gameModel.getPaddleAY()+gameModel.getPaddleHeight(),white);
-            c.drawRect(gameModel.getScreenWidth()-gameModel.getPaddleWidth(),gameModel.getPaddleBY(), gameModel.getScreenWidth(),gameModel.getPaddleBY()+gameModel.getPaddleHeight(),white);
-            c.drawCircle(gameModel.getBallX()+gameModel.getDiameter()/2,gameModel.getBallY()+gameModel.getDiameter()/2,gameModel.getDiameter()/2,white);
-            holder.unlockCanvasAndPost(c);
+    public void draw(){
+        if(!holder.getSurface().isValid()){
+            return;
         }
+        Paint white = new Paint();
+        white.setStyle(Paint.Style.FILL);
+        white.setColor(Color.WHITE);
+        Canvas c = holder.lockCanvas();
+        c.drawARGB(255,0,0,0);
+        c.drawRect(0,gameModel.getPaddleAY(), gameModel.getPaddleWidth(),gameModel.getPaddleAY()+gameModel.getPaddleHeight(),white);
+        c.drawRect(gameModel.getScreenWidth()-gameModel.getPaddleWidth(),gameModel.getPaddleBY(), gameModel.getScreenWidth(),gameModel.getPaddleBY()+gameModel.getPaddleHeight(),white);
+        c.drawCircle(gameModel.getBallX()+gameModel.getDiameter()/2,gameModel.getBallY()+gameModel.getDiameter()/2,gameModel.getDiameter()/2,white);
+        holder.unlockCanvasAndPost(c);
     }
 
-    public void pause(){
-        isRunning=false;
-        while(true){
-            try {
-                gameThread.join();
-            } catch (InterruptedException e) {
-                e.printStackTrace();
-            }
-            break;
-        }
-        gameThread=null;
-    }
-
-    public void resume(){
-        isRunning=true;
-        gameThread = new Thread(this);
-        gameThread.start();
-    }
 
 
 }
